finally got rid of the json files
The pain of json hand-writing is over now that data files are written in
yaml. The bonus is that json files are also minified too. It makes more
requirements for building the presentation, but this really is such a
tiny price :)

The `install-check` target was renamed to `env-check` since the former
felt a bit confusing.