EarthWalker could also be called the MoonWalker
A new Journey class was spun off Earth for better separation of concerns
but as a side effect it made it easier to implement backward traveling.

A few things needed to be changed like giving a direction in which we
cross the border of the map instead of just saying that we cross it.